# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Peak Academy API
  version: 1.0.0
servers:
  - url: http://localhost:4000
paths:
...
  /api/sessions:
    get:
      summary: List sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sessions list
    post:
      summary: Create teacher session
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Session created
...
  /api/earnings:
    get:
      summary: Teacher earnings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Earnings list
...
  /api/questions:
    post:
      summary: Ask question
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Question created
    get:
      summary: List unanswered questions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Questions list
...
  /api/admin/stats:
    get:
      summary: Admin stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats data
  /api/admin/users:
    get:
      summary: Admin users list
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Users list
...
  /api/notifications:
    get:
      summary: List notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications list
    post:
      summary: Create notification
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notification created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT